Move inner guts to the pio-core crate.
