+++ pio-core/README.md
@@ -1,0 +1,4 @@
+# pio-core
+
+This crate is an implementation detail, you must not use it directly.
+Use the `pio` crate instead.
