+++ pio-parser/src/pio.lalrpop
@@ -1,7 +1,7 @@
 // https://github.com/raspberrypi/pico-sdk/blob/master/tools/pioasm/parser.yy
 
 use std::str::FromStr;
-use ::pio::{
+use ::pio_core::{
   JmpCondition,
   WaitSource,
   InSource,
